requirements.txt
aiofiles==24.1.0
allure-pytest==2.15.0
allure-python-commons==2.15.0
anyio==4.10.0
async-property==0.2.2
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
deprecation==2.1.0
gherkin-official==29.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jwcrypto==1.5.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
parse==1.20.2
parse_type==0.6.4
playwright==1.54.0
pluggy==1.6.0
pycparser==2.22
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-bdd==8.1.0
python-keycloak==5.7.0
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.1
urllib3==2.5.0
